+++ build.gradle
@@ -37,7 +37,7 @@ dependencies {
 	// PostgreSQL JDBC 드라이버 의존성
 	runtimeOnly 'org.postgresql:postgresql'
 	// h2
-//	runtimeOnly 'com.h2database:h2'
+	runtimeOnly 'com.h2database:h2'
 
 
 	// Jakarta Validation API 의존성
