+++ src/main/java/com/chukapoka/server/user/sevice/UserService.java
@@ -167,6 +167,7 @@ public class UserService {
         );
     }
 
+
     /** 유저정보가 담긴 Authentication으로 토큰 생성 및 Token 테이블에 등록 */
     public TokenResponseDto saveToken(Authentication authentication){
         // JWT 토큰 생성
